backend/requirements.txt
astroid==2.2.5
click==8.1.7
ecdsa==0.13.2
Flask==2.2.5
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.0.5
future==0.17.1
greenlet==3.0.3
importlib-metadata==6.7.0
isort==4.3.18
itsdangerous==2.1.2
Jinja2==3.1.4
lazy-object-proxy==1.4.0
MarkupSafe==2.1.5
mccabe==0.6.1
pycryptodome==3.3.1
pylint==2.3.1
python-dotenv==0.21.1
python-jose-cryptodome==1.3.2
six==1.12.0
SQLAlchemy==2.0.31
typed-ast==1.4.2
typing_extensions==4.7.1
Werkzeug==2.2.3
wrapt==1.11.1
zipp==3.15.0
